Add a work-in-progress attempt to port to cygwin + windows python.
This should *not* be used in production yet, and includes a reasonable
amount of debug stuff.

The changes the sitepackage are mostly complete, but the changes in the
wrapper are highly experimental. We probably need a better (global)
way to detect that we're on cygwin, and then behave differently in
a few key places.

The files in the new cygwin folder should either be copied to somewhere
that's on your $PATH, or just add that folder to $PATH.