Add a script which imports a repo of tasks into trac.
Said repo contains a collection of yaml files, each representing
a single task which will become a ticket. This builds on our existing
trac support, so includes dependency resolution before addition.

Each task can be uniquely referred to through use of its task name. This
is the path to the file relative to the root of the repo and without the
'.yaml' extension. For example: 'comp/main'.

Attributes on the tickets are represented by top level keys within the files,
with the addition of a 'dependencies' key.
This allows the listing of task dependencies in the following ways:
 * By the name of the task
 * By a glob on the name of some tasks
 * By existing ticket number within the target trac instance, prefixed
   with an underscore rather than a hash so that it doesn't get treated
   as a comment by the yaml loader. For example: '_545'.

Change-Id: I504aae5ca1ec75199ff04dedb3a3a98ce1c50585